Firebase messaging integrated with irf creation
-Got rid of FirebaseMessageService, now using functions
-Created FirebaseAppConfig to handle wiring of signals
-Firebase code now in notifications.py in firebase app
